Just some type-conversions
git-svn-id: https://bwapi.googlecode.com/svn/trunk@450 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAI/Source/Formula/Function.cpp
@@ -64,9 +64,9 @@ namespace Formula
   {
     switch (this->type)
     {
-      case FunctionType::Finished   : return BWAPI::Broodwar.BWAPIPlayer->getCompletedUnits(this->unitType.getID());
-      case FunctionType::All        : return BWAPI::Broodwar.BWAPIPlayer->getAllUnits(this->unitType.getID());
-      case FunctionType::InProgress : return BWAPI::Broodwar.BWAPIPlayer->getIncompleteUnits(this->unitType.getID());
+      case FunctionType::Finished   : return (float) BWAPI::Broodwar.BWAPIPlayer->getCompletedUnits(this->unitType.getID());
+      case FunctionType::All        : return (float) BWAPI::Broodwar.BWAPIPlayer->getAllUnits(this->unitType.getID());
+      case FunctionType::InProgress : return (float) BWAPI::Broodwar.BWAPIPlayer->getIncompleteUnits(this->unitType.getID());
       case FunctionType::ConfigValue: 
         return this->configValue;
       case FunctionType::ActiveMinerals :
